refactor: move validator to validators.go
