+++ internal/provider/dedicatedserver/validators.go
@@ -1,4 +1,4 @@
-package validator
+package dedicatedserver
 
 import (
 	"context"
@@ -8,10 +8,14 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 )
 
-// GreaterThanZeroValidator ensures that the given value is greater than zero.
-type GreaterThanZeroValidator struct{}
+// greaterThanZeroValidator ensures that the given value is greater than zero.
+type greaterThanZeroValidator struct{}
 
-func (v GreaterThanZeroValidator) ValidateString(ctx context.Context, request validator.StringRequest, response *validator.StringResponse) {
+func (v greaterThanZeroValidator) ValidateString(
+	_ context.Context,
+	request validator.StringRequest,
+	response *validator.StringResponse,
+) {
 
 	if intValue, err := strconv.Atoi(request.ConfigValue.ValueString()); err == nil {
 		if intValue <= 0 {
@@ -36,17 +40,17 @@ func (v GreaterThanZeroValidator) ValidateString(ctx context.Context, request va
 	response.Diagnostics.AddError("Invalid Value", "The value is not a valid number")
 }
 
-var _ validator.String = GreaterThanZeroValidator{}
+var _ validator.String = greaterThanZeroValidator{}
 
-func (v GreaterThanZeroValidator) Description(ctx context.Context) string {
+func (v greaterThanZeroValidator) Description(_ context.Context) string {
 	return "Ensures that the value is greater than 0"
 }
 
-func (v GreaterThanZeroValidator) MarkdownDescription(ctx context.Context) string {
+func (v greaterThanZeroValidator) MarkdownDescription(ctx context.Context) string {
 	return v.Description(ctx)
 }
 
-// GreaterThanZero returns a new instance of the validator.
-func GreaterThanZero() validator.String {
-	return GreaterThanZeroValidator{}
+// greaterThanZero returns a new instance of the validator.
+func greaterThanZero() validator.String {
+	return greaterThanZeroValidator{}
 }
